#36@major: upgrade mdx-plugin-detect-imports to ESM and use acorn to generate the proper AST + estree-util-visit to offset all nodes
