+++ packages/mdx-plugin-detect-imports/src/plugin.cjs
@@ -1,0 +1,7 @@
+module.exports =
+    ({ exportName = "importStatements" } = {}) =>
+    async (tree, file) => {
+        const getPlugin = await import("./plugin.mjs");
+        const configuredPlugin = getPlugin.default({ exportName });
+        await configuredPlugin(tree, file);
+    };
